(v2.11) Creation of Starwalkers's packahe Minor fix
