+++ starwalkers/language/fr.py
