+++ starwalkers/settings.json
@@ -1,5 +1,5 @@
 {
-  "version": "2.1d",
+  "version": "2.11",
   "starting_money": 100,
   "cost_case": 10,
   "ship_fleet": 10,
